requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.2
filelock==3.18.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2025.7.0
gunicorn==21.2.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nltk==3.8.1
numpy==1.26.4
packaging==25.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
python-docx==1.1.0
python-magic==0.4.27
PyYAML==6.0.2
regex==2025.7.34
requests==2.31.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.16.1
SQLAlchemy==2.0.42
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.1.1
tqdm==4.67.1
transformers==4.35.2
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.0.1
WTForms==3.1.0
